+++ src/app/home/home.module.ts
@@ -5,15 +5,14 @@ import { FormsModule } from '@angular/forms';
 import { HomePage } from './home.page';
 
 import { HomePageRoutingModule } from './home-routing.module';
-import { CameraComponent } from './camera/camera.component';
-import { PhotoComponent } from './photo/photo.component';
-import { LoadingComponent } from './loading/loading.component';
-import { ModalCloseButtonComponentModule } from "../shared/components/modal-close-button/modal-close-button-component.module";
-import { CardComponent } from './card/card.component';
-import { PreviewComponent } from './card/preview/preview.component';
-import { ReportComponent } from "./report/report.component";
+import { ModalCloseButtonComponent } from '../shared/components/modal-close-button/modal-close-button.component';
 
 import { SettingsModule } from '../shared/components/settings/settings.module';
+import { PhotoComponent } from '../shared/components/photo/photo.component';
+import { CameraComponent } from '../shared/components/camera/camera.component';
+import { LoadingComponent } from '../shared/components/loading/loading.component';
+import { PreviewComponent } from '../shared/components/preview/preview.component';
+import { CardComponent } from '../shared/components/card/card.component';
 
 
 @NgModule({
@@ -23,16 +22,15 @@ import { SettingsModule } from '../shared/components/settings/settings.module';
     IonicModule,
     HomePageRoutingModule,
     SettingsModule,
-    ModalCloseButtonComponentModule,
+    ModalCloseButtonComponent,
+    PhotoComponent,
+    CameraComponent,
+    LoadingComponent,
+    PreviewComponent,
+    CardComponent,
   ],
     declarations: [
       HomePage,
-      CameraComponent,
-      PhotoComponent,
-      LoadingComponent,
-      CardComponent,
-      PreviewComponent,
-      ReportComponent,
     ],
 })
 export class HomePageModule {}