+++ src/app/shared/components/modal-close-button/modal-close-button.component.ts
@@ -1,17 +1,18 @@
-import { Component, OnInit } from '@angular/core';
-import {ModalController} from '@ionic/angular';
+import { Component } from '@angular/core';
+import { IonicModule, ModalController } from '@ionic/angular';
 
 @Component({
   selector: 'app-modal-close-button',
   templateUrl: './modal-close-button.component.html',
   styleUrls: ['./modal-close-button.component.scss'],
+  standalone: true,
+  imports: [
+    IonicModule
+  ]
 })
-export class ModalCloseButtonComponent  implements OnInit {
+export class ModalCloseButtonComponent {
 
   constructor(
     public modalController: ModalController
   ) { }
-
-  ngOnInit(): void {}
-
 }