+++ src/app/shared/types/github/GithubLabel.ts
@@ -1,0 +1,11 @@
+type GithubLabel = {
+  id: number;
+  node_id: string;
+  url: string;
+  name: string;
+  color: string;
+  default: boolean;
+  description: string;
+}
+
+export default GithubLabel;